refactor(jvspatial): split Server, unify DB factory, harden DB logging
- Remove dead Server._register_core_routes; keep AppBuilder as single core-route path
- Route DatabaseManager prime DB through create_database(env=); add _instantiate_builtin_database
- Extract Server into mixins: app factory, lifecycle hooks, registration, uvicorn run
- DBLogHandler: optional database/database_getter, sync_emit; document async test teardown
- Clarify DynamoDB index env ownership in module doc; document middleware order in api-architecture.md
- Mypy: stricter factory module, suppress mixin attr false positives; fix optional SQLite/DynamoDB stubs
- Fix unregister_walker_endpoint to await unregister_walker_class
