+++ .env.example
@@ -256,6 +256,11 @@ JVSPATIAL_DB_LOGGING_DB_NAME=logs
 # Default: true
 JVSPATIAL_DB_LOGGING_API_ENABLED=true
 
+# Default log retention period in days (optional)
+# When set, jvagent uses this when bootstrapping app.yaml if context.log_retention_days is omitted.
+# 0 = indefinite. Unset = fall back to app.yaml context or built-in default (60).
+# JVSPATIAL_LOG_RETENTION_DEFAULT_DAYS=60
+
 # Logging database type
 # Options: json, sqlite, mongodb, dynamodb
 # Default: same as JVSPATIAL_DB_TYPE
@@ -368,8 +373,8 @@ JVSPATIAL_LOG_DB_PATH=./jvspatial.logs
 # JVSPATIAL_EVENTBRIDGE_SCHEDULER_GROUP=default
 
 # Lambda Web Adapter: pass-through path for direct-invoke / EventBridge JSON → HTTP POST.
-# Default is set in-process on first load_env() for AWS serverless (typically /api/_internal/deferred).
-# Set explicitly in Lambda env if your LWA extension reads variables before Python starts.
+# Server sets AWS_LWA_* via setdefault when LWA is detected on AWS serverless; IaC is still recommended.
+# JVSPATIAL_LWA_ENV_DEFAULTS=  # true/yes/1 to force; false/0/no/off to disable
 # AWS_LWA_PASS_THROUGH_PATH=/api/_internal/deferred
 # AWS_LWA_INVOKE_MODE=RESPONSE_STREAM
