+++ .github/workflows/test-jvspatial.yaml
@@ -22,8 +22,8 @@ jobs:
       run: |
         python -m pip install --upgrade pip
 
-    - name: Install jvspatial package with test dependencies
-      run: pip install -e '.[test]'
+    - name: Install jvspatial package with test and lambda dependencies
+      run: pip install -e '.[test,lambda]'
 
     - name: Run tests
       run: pytest tests/ -v --tb=short --cov=jvspatial --cov-report=xml --cov-report=term-missing --cov-fail-under=50
