+++ docs/md/server-api.md
@@ -238,7 +238,7 @@ configure_standard_logging(level="INFO", enable_colors=True)
 - Colors: only the level name is colorized for readability; set `enable_colors=False` to disable.
 - The `Server.run()` path applies this format and passes a matching `log_config` to uvicorn so startup/access logs stay aligned. Consumers like `jvagent/cli.py` also call `configure_standard_logging` to inherit the same format.
 
-**Tip:** set `log_level` in `ServerConfig` or `JVAGENT_LOG_LEVEL` to control verbosity.
+**Tip:** set `log_level` in `ServerConfig` or `JVSPATIAL_LOG_LEVEL` to control verbosity.
 
 ## Walker Endpoints
