+++ jvspatial/api/server.py
@@ -8,59 +8,31 @@ setup, lifecycle management, and endpoint routing.
 
 import asyncio
 import logging
-from typing import (
-    Any,
-    Callable,
-    Dict,
-    List,
-    Optional,
-    Type,
-    Union,
-    cast,
-)
+from typing import Any, Callable, Dict, List, Optional, Type, Union
 
-import uvicorn
 from fastapi import FastAPI
-from fastapi.responses import JSONResponse
 
 from jvspatial.api.components import AppBuilder, EndpointManager
 from jvspatial.api.components.error_handler import APIErrorHandler
 from jvspatial.api.config import ServerConfig
-from jvspatial.api.constants import APIRoutes
 from jvspatial.api.endpoints.router import EndpointRouter
 from jvspatial.api.middleware.manager import MiddlewareManager
-from jvspatial.api.server_configurator import ServerConfigurator
+from jvspatial.api.server_app_factory import ServerAppFactoryMixin
+from jvspatial.api.server_lifecycle import ServerLifecycleMixin
+from jvspatial.api.server_registration import ServerRegistrationMixin
+from jvspatial.api.server_run import ServerRunMixin
 from jvspatial.api.services.discovery import EndpointDiscoveryService
 from jvspatial.api.services.lifecycle import LifecycleManager
 from jvspatial.core.context import GraphContext
-from jvspatial.core.entities import Node, Root, Walker
-from jvspatial.logging import configure_standard_logging
+from jvspatial.core.entities import Node
 
 
-class _LevelColorFormatter(logging.Formatter):
-    """Colorize only the level name to match jvspatial console format."""
-
-    _LEVEL_COLORS = {
-        "DEBUG": "\033[36m",  # Cyan
-        "INFO": "\033[32m",  # Green
-        "WARNING": "\033[33m",  # Yellow
-        "ERROR": "\033[31m",  # Red
-        "CRITICAL": "\033[41m\033[97m",  # White on red background
-    }
-    _RESET = "\033[0m"
-
-    def format(self, record: logging.LogRecord) -> str:  # type: ignore[override]
-        color = self._LEVEL_COLORS.get(record.levelname, "")
-        original_levelname = record.levelname
-        if color:
-            record.levelname = f"{color}{record.levelname}{self._RESET}"
-        try:
-            return super().format(record)
-        finally:
-            record.levelname = original_levelname
-
-
-class Server:
+class Server(
+    ServerAppFactoryMixin,
+    ServerRegistrationMixin,
+    ServerLifecycleMixin,
+    ServerRunMixin,
+):
     """Base server class for FastAPI applications using jvspatial.
 
     This class provides core server functionality including:
@@ -98,7 +70,7 @@ class Server:
     """
 
     def __init__(
-        self: "Server",
+        self,
         config: Optional[Union[ServerConfig, Dict[str, Any]]] = None,
         node_types: Optional[List[Type[Node]]] = None,
         on_startup: Optional[List[Callable]] = None,
@@ -127,7 +99,6 @@ class Server:
                 when a password reset is requested. Use to send reset email.
             **kwargs: Additional configuration parameters forwarded to ServerConfig
         """
-        # Extract callbacks that are not part of ServerConfig
         self._on_admin_bootstrapped = on_admin_bootstrapped
         self._on_user_registered = on_user_registered
         self._on_enrich_current_user = on_enrich_current_user
@@ -144,16 +115,18 @@ class Server:
             )
         }
 
-        # Initialize configuration using clean merging
         merged_config = self._merge_config(config, config_kwargs)
 
         self.config = ServerConfig(**merged_config)
 
-        from jvspatial.runtime.lwa import apply_aws_lwa_env_defaults
+        from jvspatial.runtime.lwa import (
+            apply_aws_eventbridge_env_default,
+            apply_aws_lwa_env_defaults,
+        )
 
+        apply_aws_eventbridge_env_default(self.config)
         apply_aws_lwa_env_defaults(self.config)
 
-        # Initialize focused components
         self.app_builder = AppBuilder(self.config)
         self.endpoint_manager = EndpointManager()
         self.error_handler = APIErrorHandler()
@@ -161,40 +134,33 @@ class Server:
         self.lifecycle_manager = LifecycleManager(self)
         self.discovery_service = EndpointDiscoveryService(self)
 
-        # Initialize application components
         self.app: Optional[FastAPI] = None
-        self.endpoint_router = EndpointRouter()  # Main router for all endpoints
+        self.endpoint_router = EndpointRouter()
         self._exception_handlers: Dict[Union[int, Type[Exception]], Callable] = {}
         self._logger = logging.getLogger(__name__)
 
         self._graph_context: Optional[GraphContext] = None
 
-        # File storage components
         self._file_interface: Optional[Any] = None
         self._proxy_manager: Optional[Any] = None
         self._file_storage_service: Optional[Any] = None
 
-        # Endpoint registry service - central tracking for all endpoints
         self._endpoint_registry = self.endpoint_manager.get_registry()
 
-        # Dynamic registration support
         self._is_running = False
         self._dynamic_routes_registered = False
-        self._app_needs_rebuild = False  # Flag to track when app needs rebuilding
-        self._has_auth_endpoints = False  # Flag to track if auth endpoints exist
+        self._app_needs_rebuild = False
+        self._has_auth_endpoints = False
         self._custom_routes: List[Dict[str, Any]] = []
 
-        # Authentication configuration
         self._auth_config: Optional[Any] = None
         self._auth_endpoints_registered = False
         self._auth_service: Optional[Any] = None
 
-        # Declarative node type registration
         if node_types:
             for node_cls in node_types:
                 self.add_node_type(node_cls)
 
-        # Declarative lifecycle hooks
         if on_startup:
             for hook in on_startup:
                 self.lifecycle_manager.add_startup_hook(hook)
@@ -202,13 +168,10 @@ class Server:
             for hook in on_shutdown:
                 self.lifecycle_manager.add_shutdown_hook(hook)
 
-        # Automatically set this server as the current server in context
-        # The most recently instantiated Server becomes the current one
         from jvspatial.api.context import set_current_server
 
         set_current_server(self)
 
-        # Flush any deferred endpoints that were decorated before server initialization
         from jvspatial.api.decorators.deferred_registry import flush_deferred_endpoints
 
         flushed_count = flush_deferred_endpoints(self)
@@ -217,18 +180,15 @@ class Server:
                 f"Registered {flushed_count} deferred endpoint(s) during server initialization"
             )
 
-        # Initialize GraphContext if database configuration is provided
         if self.config.database.db_type:
             self._initialize_graph_context()
 
-        # Configure authentication if enabled (after context is initialized)
         self._configure_authentication()
 
-        # Initialize file storage if enabled
         if self.config.file_storage.file_storage_enabled:
             self._initialize_file_storage()
 
-    async def _bootstrap_admin_startup(self: "Server") -> None:
+    async def _bootstrap_admin_startup(self) -> None:
         """Create admin user on startup when bootstrap_admin_* config is set."""
         import os
 
@@ -266,17 +226,15 @@ class Server:
         except Exception as e:
             self._logger.exception("Failed to bootstrap admin: %s", e)
 
-    def _configure_authentication(self: "Server") -> None:
+    def _configure_authentication(self) -> None:
         """Configure authentication middleware and register auth endpoints if enabled."""
         from jvspatial.api.components.auth_configurator import AuthConfigurator
 
-        # Validate configuration
         if (
             self.config.auth.auth_enabled
             and self.config.auth.jwt_secret
             == "your-secret-key"  # pragma: allowlist secret
         ):
-            # Warn if using default JWT secret in production
             self._logger.warning(
                 "⚠️  Using default JWT secret. Set jwt_secret in production!"
             )
@@ -288,107 +246,12 @@ class Server:
         self._has_auth_endpoints = auth_configurator.has_auth_endpoints
         self._auth_service = auth_configurator.auth_service
 
-        # Register bootstrap admin startup hook when configured
         if (
             self._auth_service
             and self.config.auth.bootstrap_admin_email
             and self.config.auth.bootstrap_admin_password
         ):
             self.lifecycle_manager.add_startup_hook(self._bootstrap_admin_startup)
-
-    def disable_auth_endpoint(self: "Server", path: str) -> bool:
-        """Disable a specific auth endpoint by removing it from the auth router.
-
-        This method removes routes from the auth router before it's included in the app.
-        The path can be either:
-        - Relative to router prefix: "/register" (recommended)
-        - Full path including router prefix: "/auth/register"
-
-        The endpoint will be accessible at "/api/auth/register" when the router
-        is included with the API prefix.
-
-        Args:
-            path: Path to disable, either relative ("/register") or full ("/auth/register")
-
-        Returns:
-            True if endpoint was found and removed, False otherwise
-        """
-        if not hasattr(self, "_auth_router") or self._auth_router is None:
-            self._logger.debug(
-                f"Auth router not available, cannot disable endpoint: {path}"
-            )
-            return False
-
-        # Normalize path (ensure it starts with /)
-        if not path.startswith("/"):
-            path = f"/{path}"
-
-        # FastAPI routers with a prefix store routes with the full path including prefix
-        # The auth router has prefix="/auth", so routes are stored as "/auth/register", etc.
-        # Normalize the input path to include "/auth" prefix if not present
-        if path.startswith("/auth/"):
-            # Already has full path
-            full_path = path
-            relative_path = path[6:]  # Remove "/auth/" for logging
-        elif path.startswith("/auth"):
-            full_path = "/auth"
-            relative_path = "/"
-        else:
-            # Relative path provided, add "/auth" prefix
-            full_path = f"/auth{path}"
-            relative_path = path
-
-        # Find and remove matching routes
-        from fastapi.routing import APIRoute
-
-        routes_to_remove = []
-        available_paths = []
-
-        # Normalize full_path for comparison (remove trailing slash if present)
-        normalized_full_path = full_path.rstrip("/")
-        # Also try the relative path in case routes don't include prefix
-        normalized_relative_path = relative_path.rstrip("/")
-
-        for route in self._auth_router.routes:
-            # Only process APIRoute instances (ignore Mount, etc.)
-            if isinstance(route, APIRoute):
-                route_path = route.path
-                if route_path:
-                    available_paths.append(route_path)
-                    # Normalize paths for comparison (strip trailing slashes)
-                    normalized_route_path = route_path.rstrip("/")
-                    normalized_full = normalized_full_path
-                    normalized_rel = normalized_relative_path
-
-                    # Match against full path (routes include the router prefix)
-                    # Try both full path and relative path matching
-                    if normalized_route_path in (normalized_full, normalized_rel):
-                        routes_to_remove.append(route)
-
-        # Remove routes from router
-        if routes_to_remove:
-            for route in routes_to_remove:
-                self._auth_router.routes.remove(route)
-                route_path = route.path if isinstance(route, APIRoute) else None
-                self._logger.info(
-                    f"🔒 Disabled auth endpoint: {route_path or full_path}"
-                )
-            return True
-        else:
-            # Provide helpful error message with available paths
-            if available_paths:
-                self._logger.warning(
-                    f"Could not find {full_path} endpoint to disable. "
-                    f"Available auth endpoints: {', '.join(available_paths)}. "
-                    f"The endpoint may have already been disabled or removed."
-                )
-            else:
-                self._logger.warning(
-                    f"Could not find {full_path} endpoint to disable. "
-                    f"No auth endpoints found in router. "
-                    f"Auth endpoints may not be registered yet."
-                )
-            return False
 
     def _merge_config(self, config, kwargs) -> Dict[str, Any]:
         """Clean configuration merging.
@@ -407,25 +270,19 @@ class Server:
         else:
             return {**config, **kwargs}
 
-    def _initialize_graph_context(self: "Server") -> None:
-        """Initialize GraphContext with current database configuration.
-
-        This sets up the prime database for core persistence operations
-        (authentication, session management) and creates a GraphContext
-        that uses the current database from DatabaseManager.
-        """
+    def _initialize_graph_context(self) -> None:
+        """Initialize GraphContext with current database configuration."""
         from jvspatial.api.components.database_configurator import DatabaseConfigurator
 
         db_configurator = DatabaseConfigurator(self.config)
         self._graph_context = db_configurator.initialize_graph_context()
 
-    def _initialize_file_storage(self: "Server") -> None:
+    def _initialize_file_storage(self) -> None:
         """Initialize file storage interface and proxy manager."""
         try:
             from jvspatial.api.services.file_storage import FileStorageService
             from jvspatial.storage import create_storage, get_proxy_manager
 
-            # Initialize file interface
             if self.config.file_storage.file_storage_provider == "local":
                 self._file_interface = create_storage(
                     provider="local",
@@ -447,11 +304,9 @@ class Server:
                     f"Unsupported file storage provider: {self.config.file_storage.file_storage_provider}"
                 )
 
-            # Initialize proxy manager if enabled
             if self.config.proxy.proxy_enabled:
                 self._proxy_manager = get_proxy_manager()
 
-            # Create FileStorageService instance
             self._file_storage_service = FileStorageService(
                 file_interface=self._file_interface,
                 proxy_manager=self._proxy_manager,
@@ -466,707 +321,23 @@ class Server:
             self._logger.error(f"❌ Failed to initialize file storage: {e}")
             raise
 
-    def middleware(self: "Server", middleware_type: str = "http") -> Callable:
-        """Add middleware to the application.
-
-        Args:
-            middleware_type: Type of middleware ("http" or "websocket")
-
-        Returns:
-            Decorator function for middleware
-        """
-
-        def decorator(func: Callable) -> Callable:
-            # Store the middleware for later async registration
-            # This is a workaround for the async/sync decorator issue
-            self.middleware_manager._custom_middleware.append(
-                {"func": func, "middleware_type": middleware_type}
-            )
-
-            return func
-
-        return decorator
-
-    def exception_handler(
-        self: "Server", exc_class_or_status_code: Union[int, Type[Exception]]
-    ) -> Callable:
-        """Add exception handler.
-
-        Args:
-            exc_class_or_status_code: Exception class or HTTP status code
-
-        Returns:
-            Decorator function for exception handlers
-        """
-
-        def decorator(func: Callable) -> Callable:
-            self._exception_handlers[exc_class_or_status_code] = func
-            return func
-
-        return decorator
-
-    async def on_startup(self: "Server", func: Callable[[], Any]) -> Callable[[], Any]:
-        """Register startup task.
-
-        Args:
-            func: Startup function
-
-        Returns:
-            The original function
-        """
-        return self.lifecycle_manager.add_startup_hook(func)
-
-    async def on_shutdown(self: "Server", func: Callable[[], Any]) -> Callable[[], Any]:
-        """Register shutdown task.
-
-        Args:
-            func: Shutdown function
-
-        Returns:
-            The original function
-        """
-        return self.lifecycle_manager.add_shutdown_hook(func)
-
-    def _rebuild_app_if_needed(self: "Server") -> None:
-        """Rebuild the FastAPI app to reflect dynamic changes.
-
-        This is necessary because FastAPI doesn't support removing routes/routers
-        at runtime, so we need to recreate the entire app.
-        """
-        if not self._is_running or self.app is None:
-            return
-
-        try:
-            self._logger.info(
-                "🔄 Rebuilding FastAPI app for dynamic endpoint changes..."
-            )
-
-            # Store the old app reference (not used but kept for clarity)
-
-            # Create a new app with current configuration
-            self.app = self._create_app_instance()
-
-            # The server will need to be restarted manually or this won't take effect
-            # in a running uvicorn server, but we can at least update our internal state
-            self._logger.warning(
-                "App rebuilt internally. For changes to take effect in a running server, "
-                "you may need to restart or use a development server with reload=True"
-            )
-
-        except Exception as e:
-            self._logger.error(f"❌ Failed to rebuild app: {e}")
-
-    def _create_app_instance(self: "Server") -> FastAPI:
-        """Create FastAPI instance using the focused AppBuilder component.
-
-        Returns:
-            FastAPI: Fully configured application instance
-        """
-        # Create base app with lifespan
-        lifespan = self.lifecycle_manager.lifespan if not self._is_running else None
-        app = self.app_builder.create_app(lifespan=lifespan)
-
-        # Configure middleware using MiddlewareManager
-        self.middleware_manager.configure_all(app)
-
-        # Apply server configuration (error context, rate limit, auth, exception handlers)
-        ServerConfigurator(self).configure_all(app)
-
-        # Register core routes using AppBuilder
-        self.app_builder.register_core_routes(app, self._graph_context, server=self)
-
-        # Sync endpoints from modules decorated with @endpoint (handles uvicorn reload)
-        from jvspatial.api.decorators.deferred_registry import sync_endpoint_modules
-
-        synced = sync_endpoint_modules(self)
-        if synced > 0:
-            self._logger.info(f"Synced {synced} endpoint(s) from module tracker")
-
-        # Discover and register endpoints from pre-loaded modules BEFORE including routers
-        # This ensures all @endpoint decorated functions/classes are registered with the
-        # endpoint router before it's included in the FastAPI app
-        if self.discovery_service.enabled and not self._is_running:
-            try:
-                discovered_count = self.discovery_service.discover_and_register()
-                if discovered_count > 0:
-                    self._logger.info(f"🔍 Endpoints: {discovered_count} discovered")
-            except Exception as e:
-                self._logger.warning(f"⚠️ Endpoint discovery failed: {e}")
-
-        # Include routers (endpoint router now contains all discovered endpoints)
-        self._include_routers(app)
-
-        # Include authentication router if configured
-        if self._auth_endpoints_registered and hasattr(self, "_auth_router"):
-            app.include_router(self._auth_router, prefix=APIRoutes.PREFIX)
-
-        # Configure OpenAPI security
-        self.app_builder.configure_openapi_security(app, self._has_auth_endpoints)
-
-        return app
-
-    def _register_core_routes(self: "Server", app: FastAPI) -> None:
-        """Register core routes (health, root).
-
-        Args:
-            app: FastAPI application instance to configure
-        """
-
-        # Add default health check endpoint
-        @app.get("/health", response_model=None)
-        async def health_check() -> Union[Dict[str, Any], JSONResponse]:
-            """Health check endpoint."""
-            try:
-                # Test database connectivity - Root.get() always returns singleton
-                root = await Root.get()
-                return {
-                    "status": "healthy",
-                    "database": "connected",
-                    "root_node": root.id,
-                    "service": self.config.title,
-                    "version": self.config.version,
-                }
-            except Exception as e:
-                return JSONResponse(
-                    status_code=503,
-                    content={
-                        "status": "unhealthy",
-                        "error": str(e),
-                        "service": self.config.title,
-                        "version": self.config.version,
-                    },
-                )
-
-        # Add root endpoint
-        @app.get("/")
-        async def root_info() -> Dict[str, Any]:
-            """Root endpoint with API information."""
-            info = {
-                "service": self.config.title,
-                "description": self.config.description,
-                "version": self.config.version,
-                "docs": self.config.docs_url,
-                "health": "/health",
-            }
-            if self.config.graph_endpoint_enabled:
-                info["graph"] = "/api/graph"
-            return info
-
-    def _include_routers(self: "Server", app: FastAPI) -> None:
-        """Include endpoint routers and dynamic routers.
-
-        Args:
-            app: FastAPI application instance to configure
-        """
-        # Include the unified endpoint router with all endpoints
-        app.include_router(self.endpoint_router.router, prefix=APIRoutes.PREFIX)
-
-        # Include any dynamic routers from registry
-        for endpoint_info in self._endpoint_registry.get_dynamic_endpoints():
-            if endpoint_info.router:
-                app.include_router(endpoint_info.router.router, prefix=APIRoutes.PREFIX)
-
-    def _register_walker_dynamically(
-        self: "Server",
-        walker_class: Type[Walker],
-        path: str,
-        methods: Optional[List[str]] = None,
-        **kwargs: Any,
-    ) -> None:
-        """Register a walker endpoint dynamically while server is running.
-
-        Args:
-            walker_class: Walker class to register
-            path: URL path for the endpoint
-            methods: HTTP methods
-            **kwargs: Additional route parameters
-        """
-        if self.app is None:
-            return
-
-        try:
-            # Create a new endpoint router for the dynamic walker
-            dynamic_router = EndpointRouter()
-            dynamic_router.endpoint(path, methods, **kwargs)(walker_class)
-
-            # Register as dynamic endpoint in registry
-            endpoint_info = self._endpoint_registry.get_walker_info(walker_class)
-            if endpoint_info:
-                endpoint_info.is_dynamic = True
-                endpoint_info.router = dynamic_router
-            # Register the new router in the existing app
-            self.app.include_router(dynamic_router.router, prefix=APIRoutes.PREFIX)
-
-            # Transfer auth metadata to the FastAPI route handler
-            for route in self.app.routes:
-                if hasattr(route, "path") and path in route.path:
-                    route_handler = route.endpoint
-                    route_handler._auth_required = getattr(
-                        walker_class, "_auth_required", False
-                    )
-                    route_handler._required_permissions = getattr(
-                        walker_class, "_required_permissions", []
-                    )
-                    route_handler._required_roles = getattr(
-                        walker_class, "_required_roles", []
-                    )
-                    break
-
-            self._logger.info(
-                f"🔄 Dynamically registered walker: {walker_class.__name__} at {path}"
-            )
-
-        except Exception as e:
-            self._logger.error(
-                f"❌ Failed to dynamically register walker {walker_class.__name__}: {e}"
-            )
-
-    def register_walker_class(
-        self: "Server",
-        walker_class: Type[Walker],
-        path: str,
-        methods: Optional[List[str]] = None,
-        **kwargs: Any,
-    ) -> None:
-        """Programmatically register a walker class.
-
-        This method allows registration of walker classes without using decorators,
-        useful for dynamic registration from external packages.
-
-        Args:
-            walker_class: Walker class to register
-            path: URL path for the endpoint
-            methods: HTTP methods (default: ["POST"])
-            **kwargs: Additional route parameters
-        """
-        if self._endpoint_registry.has_walker(walker_class):
-            self._logger.warning(f"Walker {walker_class.__name__} already registered")
-            return
-
-        # Register with endpoint registry
-        self._endpoint_registry.register_walker(
-            walker_class,
-            path,
-            methods or ["POST"],
-            router=self.endpoint_router,
-            **kwargs,
-        )
-
-        if self._is_running and self.app is not None:
-            self._register_walker_dynamically(walker_class, path, methods, **kwargs)
-        else:
-            # Pre-configure walker class with endpoint for discovery
-            walker_class._jvspatial_endpoint_config = {  # type: ignore[attr-defined]
-                "path": path,
-                "methods": methods or ["POST"],
-                "kwargs": kwargs,
-            }
-            # Register with router
-            self.endpoint_router.endpoint(path, methods, **kwargs)(walker_class)
-
-        self._logger.info(
-            f"📝 Registered walker class: {walker_class.__name__} at {path}"
-        )
-
-    async def unregister_walker_class(
-        self: "Server", walker_class: Type[Walker]
-    ) -> bool:
-        """Remove a walker class and its endpoint from the server.
-
-        Args:
-            walker_class: Walker class to remove
-
-        Returns:
-            True if the walker was successfully removed, False otherwise
-        """
-        if not self._endpoint_registry.has_walker(walker_class):
-            self._logger.warning(f"Walker {walker_class.__name__} not registered")
-            return False
-
-        try:
-            # Unregister from endpoint registry
-            success = self._endpoint_registry.unregister_walker(walker_class)
-
-            if not success:
-                return False
-
-            # Mark app for rebuilding if server is running
-            if self._is_running:
-                self._app_needs_rebuild = True
-                self._rebuild_app_if_needed()
-                self._logger.info(
-                    f"🔄 FastAPI app rebuilt to remove walker endpoint: {walker_class.__name__}"
-                )
-
-            self._logger.info(f"🗑️ Unregistered walker class: {walker_class.__name__}")
-            return True
-
-        except Exception as e:
-            self._logger.error(
-                f"❌ Failed to unregister walker {walker_class.__name__}: {e}"
-            )
-            return False
-
-    async def unregister_walker_endpoint(
-        self: "Server", path: str
-    ) -> List[Type[Walker]]:
-        """Remove all walkers registered to a specific path.
-
-        Args:
-            path: The URL path to remove walkers from
-
-        Returns:
-            List of walker classes that were removed
-        """
-        removed_walkers: List[Type[Walker]] = []
-
-        # Get all endpoints at this path from registry
-        endpoints = self._endpoint_registry.get_by_path(path)
-
-        # Remove walker endpoints
-        for endpoint_info in endpoints:
-            if endpoint_info.endpoint_type.value == "walker":
-                handler = endpoint_info.handler
-                # Type check: ensure handler is a Type (class) before treating as walker
-                if isinstance(handler, type):
-                    walker_class = cast(Type[Walker], handler)
-                    if self.unregister_walker_class(walker_class):
-                        removed_walkers.append(walker_class)
-
-        if removed_walkers:
-            walker_names = [cls.__name__ for cls in removed_walkers]
-            self._logger.info(
-                f"🗑️ Removed {len(removed_walkers)} walkers from path {path}: {walker_names}"
-            )
-
-        return removed_walkers
-
-    async def unregister_endpoint(
-        self: "Server", endpoint: Union[str, Callable]
-    ) -> bool:
-        """Remove a function endpoint from the server.
-
-        Args:
-            endpoint: Either the path string or the function to remove
-
-        Returns:
-            True if the endpoint was successfully removed, False otherwise
-        """
-        if isinstance(endpoint, str):
-            # Remove by path using registry
-            path = endpoint
-            removed_count = self._endpoint_registry.unregister_by_path(path)
-
-            if removed_count > 0:
-                self._logger.info(
-                    f"🗑️ Removed {removed_count} endpoints from path {path}"
-                )
-                success = True
-            else:
-                self._logger.warning(f"No endpoints found at path {path}")
-                success = False
-
-        elif callable(endpoint):
-            # Remove by function reference
-            func = endpoint
-
-            if not self._endpoint_registry.has_function(func):
-                self._logger.warning(f"Function {func.__name__} not registered")
-                return False
-
-            # Unregister from registry
-            success = self._endpoint_registry.unregister_function(func)
-
-            if success:
-                self._logger.info(f"🗑️ Removed function endpoint: {func.__name__}")
-
-        else:
-            self._logger.error(
-                "Invalid endpoint parameter: must be string path or callable function"
-            )
-            return False
-
-        # Mark app for rebuilding if server is running and we removed something
-        if success and self._is_running:
-            self._app_needs_rebuild = True
-            self._rebuild_app_if_needed()
-            self._logger.info("🔄 FastAPI app rebuilt to remove function endpoint")
-
-        return success
-
-    async def unregister_endpoint_by_path(self: "Server", path: str) -> int:
-        """Remove all endpoints (both walker and function) from a specific path.
-
-        Args:
-            path: The URL path to remove all endpoints from
-
-        Returns:
-            Number of endpoints removed
-        """
-        # Use registry to remove all endpoints at path
-        removed_count = self._endpoint_registry.unregister_by_path(path)
-
-        if removed_count > 0:
-            self._logger.info(
-                f"🗑️ Removed {removed_count} total endpoints from path {path}"
-            )
-
-        return removed_count
-
-    async def list_function_endpoints(self: "Server") -> Dict[str, Dict[str, Any]]:
-        """Get information about all registered function endpoints.
-
-        Returns:
-            Dictionary mapping function names to their endpoint information
-        """
-        return self._endpoint_registry.list_functions()
-
-    def list_all_endpoints(self: "Server") -> Dict[str, Any]:
-        """Get information about all registered endpoints (walkers and functions).
-
-        Returns:
-            Dictionary with 'walkers' and 'functions' keys containing endpoint information
-        """
-        return self._endpoint_registry.list_all()
-
-    def list_walker_endpoints(self: "Server") -> Dict[str, Dict[str, Any]]:
-        """Get information about all registered walkers.
-
-        Returns:
-            Dictionary mapping walker class names to their endpoint information
-        """
-        return self._endpoint_registry.list_walkers()
-
-    def enable_package_discovery(
-        self: "Server", enabled: bool = True, patterns: Optional[List[str]] = None
-    ) -> None:
-        """Enable or disable automatic package discovery.
-
-        Args:
-            enabled: Whether to enable package discovery
-            patterns: List of package name patterns to search for
-        """
-        self.discovery_service.enable(enabled, patterns)
-
-    def refresh_endpoints(self: "Server") -> int:
-        """Refresh and discover new endpoints from packages.
-
-        Returns:
-            Number of new endpoints discovered
-        """
-        if not self._is_running:
-            self._logger.warning("Cannot refresh endpoints - server is not running")
-            return 0
-
-        return self.discovery_service.discover_and_register()
-
-    def _create_app(self: "Server") -> FastAPI:
-        """Create and configure the FastAPI application."""
-        return self._create_app_instance()
-
-    def get_app(self: "Server") -> FastAPI:
-        """Get the FastAPI application instance.
-
-        Returns:
-            Configured FastAPI application
-        """
-        if self.app is None:
-            self.app = self._create_app()
-        return self.app
-
-    def run(
-        self: "Server",
-        host: Optional[str] = None,
-        port: Optional[int] = None,
-        reload: Optional[bool] = None,
-        app_path: Optional[str] = None,
-        **uvicorn_kwargs: Any,
-    ) -> None:
-        """Run the server using uvicorn.
-
-        Args:
-            host: Override host address
-            port: Override port number
-            reload: Enable auto-reload for development
-            app_path: Import string for the ASGI app, e.g. ``"app.main:app"``.
-                Required when ``reload=True`` so uvicorn can re-import the
-                module in each reload worker.  When omitted and reload is
-                enabled a warning is logged and the app object is used
-                directly (reload will not work correctly).
-            **uvicorn_kwargs: Additional uvicorn parameters
-        """
-        # Set up standard logging (colorized level names, consistent format)
-        # Preserve DBLogHandler to ensure database logging continues to work
-        configure_standard_logging(
-            level=self.config.log_level,
-            enable_colors=True,
-            preserve_handler_class_names=["DBLogHandler", "StartupLogCounter"],
-        )
-
-        # Use provided values or fall back to config
-        run_host = host or self.config.host
-        run_port = port or self.config.port
-        run_reload = reload if reload is not None else self.config.debug
-
-        # Log concise server startup info
-        server_info = f"http://{run_host}:{run_port}"
-        if self.config.docs_url:
-            server_info += f" | docs: {self.config.docs_url}"
-        self._logger.info(f"🔧 Server: {server_info}")
-
-        # Configure uvicorn parameters with aligned logging format
-        formatter = _LevelColorFormatter(
-            fmt="%(asctime)s | %(levelname)s | %(name)s | %(message)s",
-            datefmt="%H:%M:%S",
-        )
-
-        uvicorn_config = {
-            "host": run_host,
-            "port": run_port,
-            "reload": run_reload,
-            "log_level": self.config.log_level,
-            "log_config": {
-                "version": 1,
-                "disable_existing_loggers": False,
-                "formatters": {
-                    "default": {
-                        "()": _LevelColorFormatter,
-                        "fmt": formatter._fmt,
-                        "datefmt": formatter.datefmt,
-                    }
-                },
-                "handlers": {
-                    "default": {
-                        "class": "logging.StreamHandler",
-                        "formatter": "default",
-                        "stream": "ext://sys.stdout",
-                    }
-                },
-                "loggers": {
-                    "uvicorn": {
-                        "handlers": ["default"],
-                        "level": self.config.log_level.upper(),
-                        "propagate": False,
-                    },
-                    "uvicorn.error": {
-                        "handlers": ["default"],
-                        "level": self.config.log_level.upper(),
-                        "propagate": False,
-                    },
-                    "uvicorn.access": {
-                        "handlers": ["default"],
-                        "level": self.config.log_level.upper(),
-                        "propagate": False,
-                    },
-                },
-            },
-            **uvicorn_kwargs,
-        }
-
-        # When reloading, uvicorn must receive an import string, not an app object.
-        # If app_path is provided and reload is active, pass the string so workers
-        # can re-import the module and re-register endpoints correctly.
-        if run_reload and app_path:
-            uvicorn.run(app_path, **uvicorn_config)
-        else:
-            if run_reload and not app_path:
-                self._logger.warning(
-                    "reload=True but no app_path provided; reload workers will "
-                    "not re-register endpoints correctly.  Pass "
-                    'app_path="module:app" to run() to fix this.'
-                )
-            app = self.get_app()
-            uvicorn.run(app, **uvicorn_config)
-
-    async def run_async(
-        self: "Server",
-        host: Optional[str] = None,
-        port: Optional[int] = None,
-        **uvicorn_kwargs: Any,
-    ) -> None:
-        """Run the server asynchronously.
-
-        Args:
-            host: Override host address
-            port: Override port number
-            **uvicorn_kwargs: Additional uvicorn parameters
-        """
-        run_host = host or self.config.host
-        run_port = port or self.config.port
-
-        app = self.get_app()
-
-        formatter = _LevelColorFormatter(
-            fmt="%(asctime)s | %(levelname)s | %(name)s | %(message)s",
-            datefmt="%H:%M:%S",
-        )
-
-        config = uvicorn.Config(
-            app,
-            host=run_host,
-            port=run_port,
-            log_level=self.config.log_level,
-            log_config={
-                "version": 1,
-                "disable_existing_loggers": False,
-                "formatters": {
-                    "default": {
-                        "()": _LevelColorFormatter,
-                        "fmt": formatter._fmt,
-                        "datefmt": formatter.datefmt,
-                    }
-                },
-                "handlers": {
-                    "default": {
-                        "class": "logging.StreamHandler",
-                        "formatter": "default",
-                        "stream": "ext://sys.stdout",
-                    }
-                },
-                "loggers": {
-                    "uvicorn": {
-                        "handlers": ["default"],
-                        "level": self.config.log_level.upper(),
-                        "propagate": False,
-                    },
-                    "uvicorn.error": {
-                        "handlers": ["default"],
-                        "level": self.config.log_level.upper(),
-                        "propagate": False,
-                    },
-                    "uvicorn.access": {
-                        "handlers": ["default"],
-                        "level": self.config.log_level.upper(),
-                        "propagate": False,
-                    },
-                },
-            },
-            **uvicorn_kwargs,
-        )
-        server = uvicorn.Server(config)
-        await server.serve()
-
-    def add_node_type(self: "Server", node_class: Type[Node]) -> None:
+    def add_node_type(self, node_class: Type[Node]) -> None:
         """Register a Node type for use in walkers.
 
         Args:
             node_class: Node subclass to register
         """
-        # This is mainly for documentation/organization purposes
-        # The actual registration happens automatically in jvspatial
         self._logger.info(f"Registered node type: {node_class.__name__}")
 
-    def configure_database(self: "Server", db_type: str, **db_config: Any) -> None:
+    def configure_database(self, db_type: str, **db_config: Any) -> None:
         """Configure database settings using GraphContext.
 
         Args:
             db_type: Database type ("json", "mongodb", etc.)
             **db_config: Database-specific configuration
         """
-        # Update configuration
         self.config.database.db_type = db_type
 
-        # Handle common database configurations
         if db_type == "json" and "base_path" in db_config:
             self.config.database.db_path = db_config["base_path"]
         elif db_type == "mongodb":
@@ -1177,12 +348,11 @@ class Server:
             if "database_name" in db_config:
                 self.config.database.db_database_name = db_config["database_name"]
 
-        # Initialize or re-initialize GraphContext
         self._initialize_graph_context()
 
         self._logger.info(f"🗄️ Database configured with GraphContext: {db_type}")
 
-    def get_graph_context(self: "Server") -> Optional[GraphContext]:
+    def get_graph_context(self) -> Optional[GraphContext]:
         """Get the GraphContext instance used by the server.
 
         Returns:
@@ -1201,7 +371,7 @@ class Server:
         """
         return self._endpoint_registry.has_path(path)
 
-    def set_graph_context(self: "Server", context: GraphContext) -> None:
+    def set_graph_context(self, context: GraphContext) -> None:
         """Set a custom GraphContext for the server.
 
         Args:
@@ -1210,23 +380,7 @@ class Server:
         self._graph_context = context
         self._logger.info("🎯 Custom GraphContext set for server")
 
-    def endpoint(
-        self, path: str, methods: Optional[List[str]] = None, **kwargs: Any
-    ) -> Callable:
-        """Endpoint decorator for the server instance.
 
-        Args:
-            path: URL path for the endpoint
-            methods: HTTP methods (default: ["POST"] for walkers, ["GET"] for functions)
-            **kwargs: Additional route parameters
-
-        Returns:
-            Decorator function for endpoints
-        """
-        return self.endpoint_manager.register_endpoint(path, methods, **kwargs)
-
-
-# Convenience function for quick server creation
 def create_server(
     title: str = "jvspatial API",
     description: str = "API built with jvspatial framework",
