+++ jvspatial/db/factory.py
@@ -8,6 +8,8 @@ Supports custom database registration for seamless extension.
 
 from typing import Any, Callable, Dict, Optional
 
+from jvspatial.env import EnvConfig, load_env
+
 from .database import Database
 from .jsondb import JsonDB
 from .manager import get_database_manager
@@ -17,7 +19,7 @@ try:  # Optional dependency (requires aiosqlite)
 
     _SQLITE_AVAILABLE = True
 except ImportError:  # pragma: no cover - dependency missing
-    SQLiteDB = None  # type: ignore[misc]
+    SQLiteDB = None  # type: ignore[misc,assignment]
     _SQLITE_AVAILABLE = False
 
 try:  # Optional dependency (requires aioboto3)
@@ -25,7 +27,7 @@ try:  # Optional dependency (requires aioboto3)
 
     _DYNAMODB_AVAILABLE = True
 except ImportError:  # pragma: no cover - dependency missing
-    DynamoDB = None  # type: ignore[misc]
+    DynamoDB = None  # type: ignore[misc,assignment]
     _DYNAMODB_AVAILABLE = False
 
 # Registry for custom database implementations
@@ -119,10 +121,81 @@ def list_database_types() -> Dict[str, str]:
     return types
 
 
+def _instantiate_builtin_database(
+    env: EnvConfig,
+    db_type: str,
+    **kwargs: Any,
+) -> Database:
+    """Create a built-in database backend from ``env`` defaults merged with ``kwargs``."""
+    if db_type == "json":
+        base_path = kwargs.get("base_path") or kwargs.get("db_path") or env.jsondb_path
+        return JsonDB(str(base_path))
+
+    if db_type == "mongodb":
+        from .mongodb import MongoDB
+
+        kw = dict(kwargs)
+        if "uri" not in kw:
+            kw["uri"] = kw.pop("connection_string", None) or kw.pop(
+                "db_connection_string", None
+            )
+        if "db_name" not in kw:
+            kw["db_name"] = kw.pop("database_name", None)
+        if "uri" not in kw:
+            kw["uri"] = env.mongodb_uri
+        if "db_name" not in kw:
+            kw["db_name"] = env.mongodb_db_name
+        if "max_pool_size" not in kw:
+            kw["max_pool_size"] = env.mongodb_max_pool
+        if "min_pool_size" not in kw:
+            kw["min_pool_size"] = env.mongodb_min_pool
+
+        return MongoDB(**kw)
+
+    if db_type == "sqlite":
+        if not _SQLITE_AVAILABLE:
+            raise ImportError(
+                "aiosqlite is required for SQLite support. Install it with: pip install aiosqlite"
+            )
+
+        sqlite_kwargs = dict(kwargs)
+        db_path = (
+            sqlite_kwargs.pop("db_path", None)
+            or sqlite_kwargs.pop("path", None)
+            or env.db_path
+        )
+        return SQLiteDB(db_path=db_path, **sqlite_kwargs)
+
+    if db_type == "dynamodb":
+        if not _DYNAMODB_AVAILABLE:
+            raise ImportError(
+                "aioboto3 is required for DynamoDB support. Install it with: pip install jvspatial[lambda]"
+            )
+
+        from .dynamodb import DynamoDB
+
+        dynamodb_kwargs = dict(kwargs)
+        if "table_name" not in dynamodb_kwargs:
+            dynamodb_kwargs["table_name"] = env.dynamodb_table_name
+        if "region_name" not in dynamodb_kwargs:
+            dynamodb_kwargs["region_name"] = env.dynamodb_region
+        if "endpoint_url" not in dynamodb_kwargs:
+            dynamodb_kwargs["endpoint_url"] = env.dynamodb_endpoint_url
+        if "aws_access_key_id" not in dynamodb_kwargs:
+            dynamodb_kwargs["aws_access_key_id"] = env.dynamodb_access_key_id
+        if "aws_secret_access_key" not in dynamodb_kwargs:
+            dynamodb_kwargs["aws_secret_access_key"] = env.dynamodb_secret_access_key
+
+        return DynamoDB(**dynamodb_kwargs)
+
+    raise ValueError(f"Not a built-in database type: {db_type!r}")
+
+
 def create_database(
     db_type: str = "json",
     register: bool = False,
     name: Optional[str] = None,
+    env: Optional[EnvConfig] = None,
     **kwargs: Any,
 ) -> Database:
     """Create a database instance with direct instantiation.
@@ -134,6 +207,8 @@ def create_database(
         db_type: Database type ('json', 'mongodb', 'sqlite', 'dynamodb', or a registered custom type)
         register: If True, register the database with DatabaseManager
         name: Database name for registration (required if register=True)
+        env: Optional :class:`~jvspatial.env.EnvConfig` to use for defaults. When omitted,
+            :func:`~jvspatial.env.load_env` is used (cached).
         **kwargs: Database-specific configuration passed to the database constructor
 
     Returns:
@@ -166,81 +241,14 @@ def create_database(
     Raises:
         ValueError: If db_type is not supported or registration fails
     """
-    from jvspatial.env import load_env
+    resolved_env = env if env is not None else load_env()
 
-    env = load_env()
-
-    # Check built-in types first
     db: Database
-    if db_type == "json":
-        # Accept base_path or db_path
-        base_path = kwargs.get("base_path") or kwargs.get("db_path") or env.jsondb_path
-        db = JsonDB(str(base_path))
-
-    elif db_type == "mongodb":
-        from .mongodb import MongoDB
-
-        # Normalize aliases: connection_string/db_connection_string -> uri, database_name -> db_name
-        if "uri" not in kwargs:
-            kwargs["uri"] = kwargs.pop("connection_string", None) or kwargs.pop(
-                "db_connection_string", None
-            )
-        if "db_name" not in kwargs:
-            kwargs["db_name"] = kwargs.pop("database_name", None)
-        # Provide defaults from environment
-        if "uri" not in kwargs:
-            kwargs["uri"] = env.mongodb_uri
-        if "db_name" not in kwargs:
-            kwargs["db_name"] = env.mongodb_db_name
-        if "max_pool_size" not in kwargs:
-            kwargs["max_pool_size"] = env.mongodb_max_pool
-        if "min_pool_size" not in kwargs:
-            kwargs["min_pool_size"] = env.mongodb_min_pool
-
-        db = MongoDB(**kwargs)
-
-    elif db_type == "sqlite":
-        if not _SQLITE_AVAILABLE:
-            raise ImportError(
-                "aiosqlite is required for SQLite support. Install it with: pip install aiosqlite"
-            )
-
-        sqlite_kwargs = kwargs.copy()
-        db_path = (
-            sqlite_kwargs.pop("db_path", None)
-            or sqlite_kwargs.pop("path", None)
-            or env.db_path
-        )
-        db = SQLiteDB(db_path=db_path, **sqlite_kwargs)
-
-    elif db_type == "dynamodb":
-        if not _DYNAMODB_AVAILABLE:
-            raise ImportError(
-                "aioboto3 is required for DynamoDB support. Install it with: pip install aioboto3"
-            )
-
-        from .dynamodb import DynamoDB
-
-        # Provide defaults from environment
-        dynamodb_kwargs = kwargs.copy()
-        if "table_name" not in dynamodb_kwargs:
-            dynamodb_kwargs["table_name"] = env.dynamodb_table_name
-        if "region_name" not in dynamodb_kwargs:
-            dynamodb_kwargs["region_name"] = env.dynamodb_region
-        if "endpoint_url" not in dynamodb_kwargs:
-            dynamodb_kwargs["endpoint_url"] = env.dynamodb_endpoint_url
-        if "aws_access_key_id" not in dynamodb_kwargs:
-            dynamodb_kwargs["aws_access_key_id"] = env.dynamodb_access_key_id
-        if "aws_secret_access_key" not in dynamodb_kwargs:
-            dynamodb_kwargs["aws_secret_access_key"] = env.dynamodb_secret_access_key
-
-        db = DynamoDB(**dynamodb_kwargs)
-
-    # Check registered custom types
+    if db_type in ("json", "mongodb", "sqlite", "dynamodb"):
+        db = _instantiate_builtin_database(resolved_env, db_type, **kwargs)
     elif db_type in _DATABASE_REGISTRY:
         factory = _DATABASE_REGISTRY[db_type]
         db = factory(**kwargs)
-
     else:
         available = ", ".join(list_database_types().keys())
         raise ValueError(
