+++ jvspatial/logging/handler.py
@@ -12,10 +12,11 @@ import sys
 import threading
 import traceback
 from datetime import datetime, timezone
-from typing import Any, Dict, Optional, Set
+from typing import Any, Callable, Dict, Optional, Set
 
 from jvspatial.core.context import GraphContext
 from jvspatial.db import get_database_manager
+from jvspatial.db.database import Database
 from jvspatial.logging.models import DBLog
 from jvspatial.runtime.serverless import is_serverless_mode
 
@@ -77,6 +78,10 @@ class DBLogHandler(logging.Handler):
         Custom log levels are fully supported. Use add_custom_log_level() to register
         custom levels, then include them in the log_levels parameter when creating
         the handler.
+
+        **Tests:** In async tests, ``emit`` may schedule ``asyncio.create_task`` for the
+        save coroutine; await pending tasks (or use ``sync_emit=True``) before tearing
+        down the event loop to avoid "Task was destroyed but it is pending" warnings.
     """
 
     def __init__(
@@ -84,6 +89,10 @@ class DBLogHandler(logging.Handler):
         database_name: str = "logs",
         enabled: bool = True,
         log_levels: Optional[Set[int]] = None,
+        *,
+        database: Optional[Database] = None,
+        database_getter: Optional[Callable[[], Optional[Database]]] = None,
+        sync_emit: bool = False,
     ):
         """Initialize the database log handler.
 
@@ -91,6 +100,12 @@ class DBLogHandler(logging.Handler):
             database_name: Name of the logging database (default: "logs")
             enabled: Whether the handler is enabled (default: True)
             log_levels: Set of log levels to capture (default: {ERROR, CRITICAL})
+            database: Optional fixed database instance for log persistence (skips manager
+                lookup by name when set).
+            database_getter: Optional callable returning the log database; overrides
+                manager lookup when provided (and ``database`` is not set).
+            sync_emit: If True, always persist via ``asyncio.run`` in the emitting thread
+                instead of ``create_task`` when a loop is running (useful in tests).
         """
         # Default to ERROR and CRITICAL if no levels specified
         if log_levels is None:
@@ -103,40 +118,56 @@ class DBLogHandler(logging.Handler):
         self._database_name = database_name
         self._enabled = enabled
         self._log_levels = log_levels
-        self._log_db = None
+        self._log_db = database
+        self._database_getter = database_getter
+        self._sync_emit = sync_emit
 
-    def _get_log_database(self):
+    def _get_log_database(self) -> Optional[Database]:
         """Get the logging database instance.
 
         Returns:
             Database instance or None if not available
         """
-        if self._log_db is None:
+        if self._log_db is not None:
+            return self._log_db
+        if self._database_getter is not None:
+            if self._log_db is not None:
+                return self._log_db
             try:
-                manager = get_database_manager()
-                registered_dbs = manager.list_databases()
-
-                if self._database_name in registered_dbs:
-                    self._log_db = manager.get_database(self._database_name)
-                    return self._log_db
-                else:
-                    if not getattr(self, "_warned_db_not_found", False):
-                        self._warned_db_not_found = True
-                        logger.warning(
-                            "DBLogHandler: database '%s' not found in manager. "
-                            "Registered: %s. DB logs will not be persisted.",
-                            self._database_name,
-                            list(registered_dbs.keys()),
-                        )
-                    return None
+                db = self._database_getter()
             except Exception as e:
                 if not getattr(self, "_warned_db_not_found", False):
                     self._warned_db_not_found = True
                     logger.warning(
-                        "DBLogHandler: failed to get log database: %s", e, exc_info=True
+                        "DBLogHandler: database_getter failed: %s", e, exc_info=True
                     )
                 return None
-        return self._log_db
+            if db is not None:
+                self._log_db = db
+            return db
+        try:
+            manager = get_database_manager()
+            registered_dbs = manager.list_databases()
+
+            if self._database_name in registered_dbs:
+                self._log_db = manager.get_database(self._database_name)
+                return self._log_db
+            if not getattr(self, "_warned_db_not_found", False):
+                self._warned_db_not_found = True
+                logger.warning(
+                    "DBLogHandler: database '%s' not found in manager. "
+                    "Registered: %s. DB logs will not be persisted.",
+                    self._database_name,
+                    list(registered_dbs.keys()),
+                )
+            return None
+        except Exception as e:
+            if not getattr(self, "_warned_db_not_found", False):
+                self._warned_db_not_found = True
+                logger.warning(
+                    "DBLogHandler: failed to get log database: %s", e, exc_info=True
+                )
+            return None
 
     def emit(self, record: logging.LogRecord) -> None:
         """Emit a log record to the database.
@@ -306,10 +337,9 @@ class DBLogHandler(logging.Handler):
                     )
 
             if not is_serverless_mode():
-                # Fire-and-forget async save when background tasks allowed
                 try:
                     loop = asyncio.get_event_loop()
-                    if loop.is_running():
+                    if loop.is_running() and not self._sync_emit:
                         asyncio.create_task(save_log())
                     else:
                         asyncio.run(save_log())
@@ -318,7 +348,7 @@ class DBLogHandler(logging.Handler):
                         asyncio.run(save_log())
             else:
                 # Sync save in thread in serverless mode (e.g. Lambda)
-                # Ensures logs persist regardless of JVAGENT_LOGGING_ENABLED
+                # Ensures logs persist when DB logging is initialized (jvagent gates setup via JVSPATIAL_DB_LOGGING_ENABLED)
                 def _run_sync_save():
                     asyncio.run(save_log())
 
@@ -381,6 +411,10 @@ def install_db_log_handler(
     database_name: str = "logs",
     enabled: bool = True,
     log_levels: Optional[Set[int]] = None,
+    *,
+    database: Optional[Database] = None,
+    database_getter: Optional[Callable[[], Optional[Database]]] = None,
+    sync_emit: bool = False,
 ) -> bool:
     """Install the database log handler to the root logger.
 
@@ -388,6 +422,9 @@ def install_db_log_handler(
         database_name: Name of the logging database (default: "logs")
         enabled: Whether the handler is enabled (default: True)
         log_levels: Set of log levels to capture (default: {ERROR, CRITICAL})
+        database: Optional explicit database for persistence (see :class:`DBLogHandler`).
+        database_getter: Optional callable to resolve the log database.
+        sync_emit: When True, avoid fire-and-forget tasks when a loop is running.
 
     Returns:
         True if handler was installed, False if already exists
@@ -409,6 +446,9 @@ def install_db_log_handler(
             database_name=database_name,
             enabled=enabled,
             log_levels=log_levels,
+            database=database,
+            database_getter=database_getter,
+            sync_emit=sync_emit,
         )
         root_logger.addHandler(handler)
